requirements.txt
APScheduler==3.6.3
atomicwrites==1.4.0
attrs==21.2.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.4
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
python-telegram-bot==13.9
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
toml==0.10.2
tornado==6.1
transitions==0.8.10
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7